+++ snuba/datasets/dataset.py
@@ -1,18 +1,14 @@
-from typing import Any, Mapping, Optional, Sequence, Tuple
+from typing import Mapping, Optional, Sequence
 
 from snuba.clickhouse.columns import ColumnSet
-from snuba.clickhouse.escaping import escape_identifier
+from snuba.datasets.entity import Entity
+from snuba.datasets.entities.factory import get_entity_name
 from snuba.datasets.plans.query_plan import ClickhouseQueryPlanBuilder
-from snuba.datasets.storage import Storage, WritableStorage, WritableTableStorage
-from snuba.query.extensions import QueryExtension
-from snuba.query.logical import Query
-from snuba.query.parsing import ParsingContext
-from snuba.query.processors import QueryProcessor
 from snuba.query.validation import FunctionCallValidator
-from snuba.util import parse_datetime, qualified_column
 
 
 class Dataset(object):
+    # TODO: s/dataset/entity/g
     """
     A dataset represents a data model we can run a Snuba Query on.
     A data model provides a logical schema (today it is a flat table,
@@ -46,94 +42,78 @@ class Dataset(object):
     manipulate the lower layer objects.
     """
 
+    # TODO: Still not sure if the Dataset class really needs the actual Entity object
+    # or just the name.
     def __init__(
-        self,
-        *,
-        storages: Sequence[Storage],
-        query_plan_builder: ClickhouseQueryPlanBuilder,
-        abstract_column_set: ColumnSet,
-        writable_storage: Optional[WritableStorage],
+        self, *, entities: Sequence[Entity], default_entity: Optional[Entity] = None
     ) -> None:
-        self.__storages = storages
-        self.__query_plan_builder = query_plan_builder
-        self.__abstract_column_set = abstract_column_set
-        self.__writable_storage = writable_storage
+        self.__entities: Mapping[str, Entity] = {
+            get_entity_name(entity): entity for entity in entities
+        }
 
-    def get_extensions(self) -> Mapping[str, QueryExtension]:
-        """
-        Returns the extensions for this dataset.
-        Every extension comes as an instance of QueryExtension.
-        The schema tells Snuba how to parse the query.
-        The processor actually does query processing for this extension.
-        """
-        raise NotImplementedError("dataset does not support queries")
+        # TODO: This is a convenience while we slowly migrate everything to Entities. This way
+        # every dataset can have a default entity which acts as a passthrough until we can
+        # migrate the datasets to proper entities.
+        self.__default_entity: Entity = None
+        if default_entity:
+            self.__default_entity = default_entity
+        elif len(entities) == 1:
+            self.__default_entity = entities[0]
 
-    def get_query_processors(self) -> Sequence[QueryProcessor]:
-        """
-        Returns a series of transformation functions (in the form of QueryProcessor objects)
-        that are applied to queries after parsing and before running them on the storage.
-        These are applied in sequence in the same order as they are defined and are supposed
-        to be stateless.
-        """
-        return []
+        assert self.__default_entity is not None, "Datasets must have a default entity"
 
-    def get_abstract_columnset(self) -> ColumnSet:
-        """
-        Returns the abstract query schema for this dataset. This is where Entities
-        will come into play since this method will return the structure of the
-        data model.
-        Now the data model is flat so this is just a simple ColumnSet object. With entities
-        this will be a more complex data structure that defines the schema for each entity
-        and their relations.
-        """
-        # TODO: Make this available to the dataset query processors.
-        return self.__abstract_column_set
+        # TODO: I want to maintain this API for now just to limit the scope of the code changes.
+        # All the code that references Datasets needs to be evaluated as to whether it should take
+        # entities instead, but for now keep these functions here.
+        deprecated_functions = [
+            "get_extensions",
+            "get_query_processors",
+            "get_all_storages",
+            "get_writable_storage",
+            "column_expr",
+        ]
+        for func in deprecated_functions:
+            setattr(self, func, getattr(self.__default_entity, func))
 
-    def get_query_plan_builder(self) -> ClickhouseQueryPlanBuilder:
+    def get_all_entities(self) -> Sequence[Entity]:
         """
-        Returns the component that transforms a Snuba query in a Storage query by selecting
-        the storage and provides the directions on how to run the query.
-        """
-        return self.__query_plan_builder
-
-    def get_all_storages(self) -> Sequence[Storage]:
-        """
-        Returns all storages for this dataset.
+        Returns all entities for this dataset.
         This method should be used for schema bootstrap and migrations.
         It is not supposed to be used during query processing.
         """
-        return self.__storages
+        return self.__entities.items()
 
-    def get_writable_storage(self) -> Optional[WritableTableStorage]:
-        """
-        Temporarily support getting the writable storage from a dataset.
-        Once consumers/replacers no longer reference datasets, this can be removed
-        and datasets can have more than one writable storage.
-        """
-        return self.__writable_storage
+    def get_entity(self, entity_name: Optional[str]) -> Entity:
+        if entity_name and entity_name in self.__entities:
+            return self.__entities[entity_name]
 
-    # Old methods that we are migrating away from
-    def column_expr(
-        self,
-        column_name,
-        query: Query,
-        parsing_context: ParsingContext,
-        table_alias: str = "",
-    ):
-        """
-        Return an expression for the column name. Handle special column aliases
-        that evaluate to something else.
-        """
-        return escape_identifier(qualified_column(column_name, table_alias))
+        return self.__default_entity
 
-    def process_condition(self, condition) -> Tuple[str, str, Any]:
+    def get_query_plan_builder(
+        self, entity_name: Optional[str] = ""
+    ) -> ClickhouseQueryPlanBuilder:
         """
-        Return a processed condition tuple.
-        This enables a dataset to do any parsing/transformations
-        a condition before it is added to the query.
-        """
-        return condition
+        Returns the component that transforms a Snuba query in a Storage query by selecting
+        the storage and provides the directions on how to run the query.
 
+        If the query is being executed on a single entity, that entity will be used to determine
+        the query plan. In cases such as a join, the dataset will something something and
+        then build the join query.
+        """
+        if entity_name:
+            entity = self.get_entity(entity_name)
+        else:
+            entity = self.__default_entity
+
+        return entity.get_query_plan_builder()
+
+    def get_abstract_columnset(self) -> ColumnSet:
+        """
+        This is just a wrapper to maintain the legacy Dataset interface. It should be removed.
+        """
+        return self.__default_entity.get_data_model()
+
+    # TODO: Entity or Dataset? Nothing seems to use this.
     def get_function_call_validators(self) -> Mapping[str, FunctionCallValidator]:
         """
         Provides a sequence of function expression validators for
@@ -141,69 +121,3 @@ class Dataset(object):
         calls to dataset specific functions are well formed.
         """
         return {}
-
-
-class TimeSeriesDataset(Dataset):
-    def __init__(
-        self,
-        *,
-        storages: Sequence[Storage],
-        query_plan_builder: ClickhouseQueryPlanBuilder,
-        abstract_column_set: ColumnSet,
-        writable_storage: Optional[WritableStorage],
-        time_group_columns: Mapping[str, str],
-        time_parse_columns: Sequence[str],
-    ) -> None:
-        super().__init__(
-            storages=storages,
-            query_plan_builder=query_plan_builder,
-            abstract_column_set=abstract_column_set,
-            writable_storage=writable_storage,
-        )
-        # Convenience columns that evaluate to a bucketed time. The bucketing
-        # depends on the granularity parameter.
-        # The bucketed time column names cannot be overlapping with existing
-        # schema columns
-        for bucketed_column in time_group_columns.keys():
-            assert (
-                bucketed_column not in abstract_column_set
-            ), f"Bucketed column {bucketed_column} is already defined in the schema"
-        self.__time_group_columns = time_group_columns
-        self.__time_parse_columns = time_parse_columns
-
-    def time_expr(self, column_name: str, granularity: int, table_alias: str) -> str:
-        real_column = qualified_column(column_name, table_alias)
-        template = {
-            3600: "toStartOfHour({column}, 'Universal')",
-            60: "toStartOfMinute({column}, 'Universal')",
-            86400: "toDate({column}, 'Universal')",
-        }.get(
-            granularity,
-            "toDateTime(multiply(intDiv(toUInt32({column}), {granularity}), {granularity}), 'Universal')",
-        )
-        return template.format(column=real_column, granularity=granularity)
-
-    def column_expr(
-        self,
-        column_name,
-        query: Query,
-        parsing_context: ParsingContext,
-        table_alias: str = "",
-    ):
-        # We want to permit functions here, so we need to make sure we're not trying
-        # to look up lists in the dictionary or it will fail with a type error.
-        if isinstance(column_name, str) and column_name in self.__time_group_columns:
-            real_column = self.__time_group_columns[column_name]
-            return self.time_expr(real_column, query.get_granularity(), table_alias)
-        else:
-            return super().column_expr(column_name, query, parsing_context, table_alias)
-
-    def process_condition(self, condition) -> Tuple[str, str, Any]:
-        lhs, op, lit = condition
-        if (
-            lhs in self.__time_parse_columns
-            and op in (">", "<", ">=", "<=", "=", "!=")
-            and isinstance(lit, str)
-        ):
-            lit = parse_datetime(lit)
-        return lhs, op, lit
