+++ tests/clickhouse/query_dsl/test_time_range.py
@@ -4,6 +4,8 @@ from snuba.clickhouse.query import Query as ClickhouseQuery
 from snuba.clickhouse.query_dsl.accessors import get_time_range
 from snuba.datasets.factory import get_dataset
 from snuba.query.parser import parse_query
+from snuba.query.processors.timeseries_processor import TimeSeriesProcessor
+from snuba.request.request_settings import HTTPRequestSettings
 
 
 def test_get_time_range() -> None:
@@ -23,6 +25,10 @@ def test_get_time_range() -> None:
 
     events = get_dataset("events")
     query = parse_query(body, events)
+    processors = events.get_query_processors()
+    for processor in processors:
+        if isinstance(processor, TimeSeriesProcessor):
+            processor.process_query(query, HTTPRequestSettings())
 
     from_date_ast, to_date_ast = get_time_range(ClickhouseQuery(query), "timestamp")
     assert (
