+++ tests/query/test_project_extension.py
@@ -56,7 +56,7 @@ def test_project_extension_query_processing(
 
 def test_project_extension_query_adds_rate_limits():
     extension = ProjectExtension(project_column="project_id")
-    raw_data = {"project": [2, 3]}
+    raw_data = {"project": [1, 2]}
     valid_data = validate_jsonschema(raw_data, extension.get_schema())
     query = Query({"conditions": []}, TableSource("my_table", ColumnSet([])),)
     request_settings = HTTPRequestSettings()
@@ -69,25 +69,25 @@ def test_project_extension_query_adds_rate_limits():
     assert len(rate_limits) == num_rate_limits_before_processing + 1
 
     most_recent_rate_limit = rate_limits[-1]
-    assert most_recent_rate_limit.bucket == "2"
+    assert most_recent_rate_limit.bucket == "1"
     assert most_recent_rate_limit.per_second_limit == 1000
     assert most_recent_rate_limit.concurrent_limit == 1000
 
 
 def test_project_extension_project_rate_limits_are_overridden():
     extension = ProjectExtension(project_column="project_id")
-    raw_data = {"project": [2, 3]}
+    raw_data = {"project": [3, 4]}
     valid_data = validate_jsonschema(raw_data, extension.get_schema())
     query = Query({"conditions": []}, TableSource("my_table", ColumnSet([])),)
     request_settings = HTTPRequestSettings()
-    state.set_config("project_per_second_limit_2", 5)
-    state.set_config("project_concurrent_limit_2", 10)
+    state.set_config("project_per_second_limit_3", 5)
+    state.set_config("project_concurrent_limit_3", 10)
 
     extension.get_processor().process_query(query, valid_data, request_settings)
 
     rate_limits = request_settings.get_rate_limit_params()
     most_recent_rate_limit = rate_limits[-1]
 
-    assert most_recent_rate_limit.bucket == "2"
+    assert most_recent_rate_limit.bucket == "3"
     assert most_recent_rate_limit.per_second_limit == 5
     assert most_recent_rate_limit.concurrent_limit == 10
